+++ packages/terra-modal-manager/src/actions.js
@@ -5,8 +5,8 @@ import {
   POP,
   MAXIMIZE,
   MINIMIZE,
-  REQUEST_FOCUS,
-  RELEASE_FOCUS,
+  GAIN_FOCUS,
+  LOSE_FOCUS,
 } from './actionTypes';
 
 export function open(data) {
@@ -33,10 +33,10 @@ export function minimize(data) {
   return { type: MINIMIZE, data };
 }
 
-export function requestFocus(data) {
-  return { type: REQUEST_FOCUS, data };
+export function gainFocus(data) {
+  return { type: GAIN_FOCUS, data };
 }
 
-export function releaseFocus(data) {
-  return { type: RELEASE_FOCUS, data };
+export function loseFocus(data) {
+  return { type: LOSE_FOCUS, data };
 }