+++ packages/terra-modal-manager/src/index.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import ModalManager from './ModalManager';
 
 import modalManagerReducers from './reducers';
-import { open, close, push, pop, maximize, minimize, requestFocus, releaseFocus } from './actions';
+import { open, close, push, pop, maximize, minimize, gainFocus, loseFocus } from './actions';
 
 const mapStateToProps = state => (
   (disclosureState => ({
@@ -24,8 +24,8 @@ const mapDispatchToProps = dispatch => ({
   popModal: (data) => { dispatch(pop(data)); },
   maximizeModal: (data) => { dispatch(maximize(data)); },
   minimizeModal: (data) => { dispatch(minimize(data)); },
-  requestFocus: (data) => { dispatch(requestFocus(data)); },
-  releaseFocus: (data) => { dispatch(releaseFocus(data)); },
+  gainFocus: (data) => { dispatch(gainFocus(data)); },
+  loseFocus: (data) => { dispatch(loseFocus(data)); },
 });
 
 export { mapDispatchToProps };