+++ packages/terra-modal-manager/src/reducers.js
@@ -1,4 +1,4 @@
-import { open, push, pop, maximize, minimize, defaultState, requestFocus, releaseFocus } from './shared/disclosureReducerUtils';
+import { open, push, pop, maximize, minimize, defaultState, gainFocus, loseFocus } from './shared/disclosureReducerUtils';
 
 import {
   OPEN,
@@ -7,8 +7,8 @@ import {
   POP,
   MAXIMIZE,
   MINIMIZE,
-  REQUEST_FOCUS,
-  RELEASE_FOCUS,
+  GAIN_FOCUS,
+  LOSE_FOCUS,
 } from './actionTypes';
 
 const supportedSizes = {
@@ -37,10 +37,10 @@ const modalManager = (state = defaultModalState, action) => {
       return maximize(state, action);
     case MINIMIZE:
       return minimize(state, action);
-    case REQUEST_FOCUS:
-      return requestFocus(state, action);
-    case RELEASE_FOCUS:
-      return releaseFocus(state, action);
+    case GAIN_FOCUS:
+      return gainFocus(state, action);
+    case LOSE_FOCUS:
+      return loseFocus(state, action);
     default:
       return state;
   }