+++ packages/terra-modal-manager/tests/nightwatch/DemoContainer.jsx
@@ -69,8 +69,8 @@ class DemoContainer extends React.Component {
         <br />
         <h4>id: {identifier}</h4>
         <br />
-        {app && app.releaseFocus ? <h4>Modal has released focus trap!</h4> : null }
-        {app && app.requestFocus ? <h4>Modal has trapped focus!</h4> : null }
+        {app && app.releaseFocus ? <h4>Modal has lost focus!</h4> : null }
+        {app && app.requestFocus ? <h4>Modal has gained focus!</h4> : null }
         <br />
         <button className="disclose" onClick={this.disclose}>Disclose</button>
         {app && app.dismiss ? <button className="dismiss" onClick={this.dismiss}>Dismiss</button> : null }